requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-filter==23.1
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
iso3166==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
stripe==5.3.0
tzdata==2023.3
urllib3==1.26.15
